+++ zfit/util/cache.py
@@ -126,8 +126,8 @@ class GraphCachable(ZfitGraphCachable):
         """
         if not isinstance(cacher, ZfitGraphCachable):
             raise TypeError(f"`cacher` is not a `ZfitGraphCachable` but {type(cacher)}")
-        if not cacher in self._cachers:
-            self._cachers[cacher] = None  # could we have a more useful value?
+        if cacher not in self._cachers:
+            self._cachers[cacher] = None
 
     def add_cache_deps(
         self, cache_deps: ztyping.CacherOrCachersType, allow_non_cachable: bool = True
@@ -149,9 +149,7 @@ class GraphCachable(ZfitGraphCachable):
                 cache_dep.register_cacher(self)
             elif not allow_non_cachable:
                 raise TypeError(
-                    "cache_dependent {} is not a `ZfitGraphCachable` but {}".format(
-                        cache_dep, type(cache_dep)
-                    )
+                    f"cache_dependent {cache_dep} is not a `ZfitGraphCachable` but {type(cache_dep)}"
                 )
 
     def reset_cache_self(self):
@@ -163,8 +161,6 @@ class GraphCachable(ZfitGraphCachable):
         self.reset_cache_self()
 
     def _clean_cache(self):
-        # for func_holder in self.graph_caching_methods:
-        #     func_holder.reset
         self._cache.clear()
         return
 
@@ -188,6 +184,7 @@ def invalidate_graph(func):
             run.clear_graph_cache()
         else:
             self.reset_cache(reseter=self)
+            # TODO: we could make the whole handling more precise and not just reset the whole graph
             # raise RuntimeError(f"The function {func} is not supported in graph mode as it modifies pieces that invalidate"
             #                    " the graph. If you think this should work, please open an issue on"
             #                    " https://github.com/zfit/zfit/issues/new/choose.")
@@ -264,7 +261,6 @@ class FunctionCacheHolder(GraphCachable):
             cachables, cachables_mapping
         )
         self._hash_value = hash((self.python_func, self.immutable_representation))
-        # self._hash_value = hash(self.immutable_representation)
         super().__init__()  # resets the cache
         self.add_cache_deps(cachables_all)
         self.is_valid = True  # needed to make the cache valid again
@@ -348,9 +344,6 @@ class FunctionCacheHolder(GraphCachable):
     def __eq__(self, other: object) -> bool:
         if not isinstance(other, FunctionCacheHolder):
             return False
-        # return False  # HACK
-        # return self.__hash__() == other.__hash__()  # HACK TODO
-        # return all(obj1 == obj2 for obj1, obj2 in zip(self.immutable_representation, other.immutable_representation))
         array_repr_self = self.immutable_representation
         array_repr_other = other.immutable_representation
         try:
@@ -358,9 +351,6 @@ class FunctionCacheHolder(GraphCachable):
                 (obj1 is obj2) or (obj1 == obj2)
                 for obj1, obj2 in zip_longest(array_repr_self, array_repr_other)
             )
-            # all_values = all(np.equal(array_repr_self, array_repr_other))
-            # if all_ids != all_values:
-            #     raise ValueError("Ids and values are not equal")
             return all_ids  # TODO: this isn't optimal...
         except ValueError:  # broadcasting does not work
             return False